Divided the home-screen into components and removed and optimized code
