+++ app/components/DR/ActionCards/ActionCards.js
@@ -1,4 +1,4 @@
-import { Badge, Button, Card, CardItem, Left, Right, Text } from 'native-base';
+import { Button, Card, Left, Text } from 'native-base';
 import React, { Component } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Image, View } from 'react-native';
@@ -13,12 +13,11 @@ import { GetStoreData } from '../../../helpers/General';
 import languages from '../../../locales/languages';
 import styles from './styles';
 
-const { ORANGE, GREEN, BLUE_RIBBON } = Colors;
+const { ORANGE, GREEN, BLUE_RIBBON, GRAY } = Colors;
 
 export class Feels extends Component {
   constructor(props) {
     super(props);
-
     this.state = {
       positive: false,
     };
@@ -86,147 +85,96 @@ export class Feels extends Component {
 export function Aurora({ navigation }) {
   const { t } = useTranslation();
   return (
-    <View>
-      <Card style={styles.bigCards}>
-        <View style={styles.auroraContainer}>
-          <Image
-            style={styles.auroraImage}
-            source={require('../../../assets/images/aurora_logo.png')}
-          />
-          <Text style={[styles.textHeader, { marginLeft: 8 }]}>
-            {t('label.auroraMsp_title')}
-          </Text>
-        </View>
-        <View style={styles.tester}>
-          <Left>
-            <Text style={styles.text}>{t('label.auroraMsp_description')}</Text>
-          </Left>
-          <Button
-            onPress={() => navigation.navigate('AuroraScreen')}
-            style={[
-              styles.buttons,
-              { backgroundColor: BLUE_RIBBON, marginLeft: 10 },
-            ]}>
-            <Text style={styles.buttonText}>{t('label.conversar_label')}</Text>
-          </Button>
-        </View>
-      </Card>
-    </View>
+    <Card style={styles.bigCards}>
+      <View style={styles.auroraContainer}>
+        <Image
+          style={styles.auroraImage}
+          source={require('../../../assets/images/aurora_logo.png')}
+        />
+        <Text style={[styles.textHeader, { marginLeft: 8 }]}>
+          {t('label.auroraMsp_title')}
+        </Text>
+      </View>
+      <View style={styles.tester}>
+        <Left>
+          <Text style={styles.text}>{t('label.auroraMsp_description')}</Text>
+        </Left>
+        <Button
+          onPress={() => navigation.navigate('AuroraScreen')}
+          style={[
+            styles.buttons,
+            { backgroundColor: BLUE_RIBBON, marginLeft: 10 },
+          ]}>
+          <Text style={styles.buttonText}>{t('label.conversar_label')}</Text>
+        </Button>
+      </View>
+    </Card>
   );
 }
 
 export function LocationMatch({ navigation }) {
   const { t } = useTranslation();
   return (
-    <View>
-      <Card style={styles.bigCards}>
-        <View style={styles.auroraContainer}>
-          <Icon name='search-location' color={BLUE_RIBBON} size={wp('6%')} />
-          <Text style={[styles.textHeader, { marginLeft: 8 }]}>
-            {t('label.location_match_title')}
+    <Card style={styles.bigCards}>
+      <View style={styles.auroraContainer}>
+        <Icon name='search-location' color={BLUE_RIBBON} size={wp('6%')} />
+        <Text style={[styles.textHeader, { marginLeft: 8 }]}>
+          {t('label.location_match_title')}
+        </Text>
+      </View>
+      <View style={styles.tester}>
+        <Left>
+          <Text style={styles.text}>
+            {t('label.location_match_description')}
           </Text>
-        </View>
-        <View style={styles.tester}>
-          <Left>
-            <Text style={styles.text}>
-              {t('label.location_match_description')}
-            </Text>
-          </Left>
-          <Button
-            onPress={() => navigation.navigate('Location')}
-            style={[
-              styles.buttons,
-              { backgroundColor: BLUE_RIBBON, marginLeft: 10 },
-            ]}>
-            <Text style={styles.buttonText}>
-              {t('label.location_match_button')}
-            </Text>
-          </Button>
-        </View>
-      </Card>
-    </View>
+        </Left>
+        <Button
+          onPress={() => navigation.navigate('Location')}
+          style={[
+            styles.buttons,
+            { backgroundColor: BLUE_RIBBON, marginLeft: 10 },
+          ]}>
+          <Text style={styles.buttonText}>
+            {t('label.location_match_button')}
+          </Text>
+        </Button>
+      </View>
+    </Card>
   );
 }
 
-export function Contact({ isProfile }) {
+export function Footer({ navigation }) {
   const { t } = useTranslation();
-  const buttonStyle = isProfile
-    ? {
-        alignSelf: 'center',
-        backgroundColor: '#0161f2',
-        borderRadius: 25,
-        justifyContent: 'center',
-        height: hp('6%'),
-        width: wp('27%'),
-      }
-    : styles.buttons;
-
-  const buttonValueStyle = isProfile
-    ? {
-        color: 'white',
-        fontSize: hp('2.1%'),
-        fontWeight: '500',
-        textAlign: 'center',
-        letterSpacing: 0.9,
-      }
-    : styles.buttonText;
-
   return (
-    <View>
-      <Card
-        style={[
-          styles.bigCards,
-          {
-            backgroundColor: BLUE_RIBBON,
-            height: hp('15%'),
-            justifyContent: 'center',
-          },
-        ]}>
-        <View style={styles.tester}>
-          <Left style={styles.tester}>
-            <Image
-              style={styles.image}
-              // eslint-disable-next-line global-require
-              source={require('../../../assets/images/logo_msp.png')}
-            />
-            <Left style={{ marginLeft: 15 }}>
-              <Text style={[styles.text, { color: '#fff' }]}>
-                {t('label.phone_line_label')}
-              </Text>
-              <Text style={[styles.textHeader, { color: '#fff' }]}>*464</Text>
-            </Left>
-            <Button style={[buttonStyle, { backgroundColor: '#fff' }]}>
-              <Text style={[buttonValueStyle, { color: '#0161F2' }]}>
-                {t('label.chat_label')}
-              </Text>
-            </Button>
-          </Left>
-        </View>
-      </Card>
-    </View>
-  );
-}
-
-export function Alerts() {
-  return (
-    <View>
-      <Card style={styles.bigCards}>
-        <CardItem header>
-          <Left>
-            <Icon name='bell-o' color={BLUE_RIBBON} size={30} />
-            <Text style={styles.textHeader}>Alertas</Text>
-          </Left>
-
-          <Right style={{ backgroundColor: '#fff' }}>
-            <Badge style={{ backgroundColor: BLUE_RIBBON }}>
-              <Text>9 +</Text>
-            </Badge>
-          </Right>
-        </CardItem>
-        <CardItem bordered>
-          <Text>Sustituir con al data de la API</Text>
-        </CardItem>
-      </Card>
+    <View style={styles.footer}>
+      <View style={{ margin: wp('5%') }}>
+        <Text style={[styles.text, { color: GRAY, textAlign: 'center' }]}>
+          {t('label.home_screen_bottom_text')}
+        </Text>
+      </View>
+      <Button
+        small
+        bordered
+        rounded
+        info
+        onPress={() => navigation.navigate('Sponsors')}
+        style={{
+          marginBottom: 10,
+          padding: 15,
+          alignSelf: 'center',
+        }}>
+        <Text
+          style={[
+            styles.text,
+            {
+              fontSize: 12,
+              color: Colors.BLUE_LINK,
+              textAlign: 'center',
+            },
+          ]}>
+          {t('label.sponsor_title')}
+        </Text>
+      </Button>
     </View>
   );
 }
